python-backend/requirements.txt
pymongo
Flask-PyMongo
flask-cors
flask-compress
flask-headers
gunicorn